bugfix: http client was decoupled from adapters layer
