+++ src/adapters/controllers/base/context/HttpMethod.enum.ts
@@ -1,0 +1,9 @@
+export enum HttpMethodEnum {
+  GET = "get",
+  POST = "post",
+  PUT = "put",
+  DELETE = "delete",
+  PATCH = "patch",
+  OPTIONS = "options",
+  HEAD = "head",
+}
