--- src/infrastructure/httpClient/SerializationType.ts
@@ -1,7 +1,0 @@
-export class SerializationType {
-  static JSON: "json";
-  static TEXT: "string";
-  static ARRAY_BUFFER: "arrayBuffer";
-  static BUFFER: "buffer";
-  static BLOB: "blob";
-}
